+++ perl/FSM/Composition/GeneratedChildRealizer.pm
@@ -219,6 +219,8 @@ sub _realize_generated_child ($class, %args) {
     my $child_module = FSM::Pipeline::SourceFrontend->create_fsm_module(
         raw_ast => $child_ast,
         debug_level => ($pipeline->{debug_level} // 0),
+        strict_mode => ($pipeline->{strict_mode} // 0),
+        source_label => $source_name,
     );
     my $child_intent_hir = FSM::IR::IntentHIRBuilder->build_from_fsm_module(
         fsm_module => $child_module,